requirements-dev.txt
backcall==0.2.0
click==8.0.3
decorator==5.1.0
greenlet==1.1.2
ipython==7.30.1
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
psycopg2==2.9.3
ptyprocess==0.7.0
pyaml==21.10.1
Pygments==2.11.1
PyYAML==6.0
SQLAlchemy==1.4.29
tabulate==0.8.9
traitlets==5.1.1
wcwidth==0.2.5
